functional search bar
